moved template to helpers directory
